fix statistics: do not count RELEASED tickets as sold
