+++ src/main/resources/alfio/db/HSQLDB/R__EVENTS_STATISTICS_VIEW.sql
@@ -16,6 +16,54 @@
 --
 
 drop view if exists events_statistics;
+drop view if exists ticket_category_statistics;
+
+create view ticket_category_statistics as (select
+  res.*,
+  is_expired and not_sold_tickets > 0 as is_containing_orphan_tickets,
+  stuck_count > 0 as is_containing_stuck_tickets
+from
+
+(select
+  id as ticket_category_id,
+  max_tickets,
+  bounded,
+  is_expired,
+  event_id,
+  coalesce(pending_count,0) as pending_count,
+  coalesce(checked_in_count,0) as checked_in_count,
+  coalesce(sold_tickets_count,0) as sold_tickets_count,
+  coalesce(released_count, 0) as released_count,
+  case(bounded) when false then 0 else max_tickets - coalesce(sold_tickets_count,0 )  - coalesce(checked_in_count, 0) -  coalesce(pending_count, 0) end as not_sold_tickets,
+  coalesce(stuck_count, 0) as stuck_count
+from
+
+(select max_tickets, bounded, id, event_id, expiration < now() as is_expired from ticket_category ) ticket_cat
+
+left join
+
+(select
+  sum(case(status = 'PENDING') when true then 1 else 0 end) as pending_count,
+  sum(case(status = 'RELEASED') when true then 1 else 0 end) as released_count,
+  sum(case(status = 'CHECKED_IN') when true then 1 else 0 end) checked_in_count,
+  sum(case(status in ('TO_BE_PAID', 'ACQUIRED')) when true then 1 else 0 end) as sold_tickets_count,
+  category_id
+from ticket
+inner join ticket_category tc
+on category_id = tc.id
+group by category_id
+) tickets_stats on ticket_cat.id = tickets_stats.category_id
+
+left join
+
+(select count(*) stuck_count, category_id
+  from ticket
+  inner join tickets_reservation on tickets_reservation.id = tickets_reservation_id
+  where tickets_reservation.status = 'STUCK'
+  group by category_id) stuck_count on ticket_cat.id = stuck_count.category_id
+
+
+) as res);
 
 create view events_statistics as (select
       event.id,
@@ -23,10 +71,12 @@ create view events_statistics as (select
       case(contains_unbounded_categories) when true then 0 else (select count(id) from ticket where event_id = event.id and status not in ('INVALIDATED', 'EXPIRED')) - allocated_count end as not_allocated_tickets,
       pending_count as pending_tickets,
       sold_tickets_count as sold_tickets,
+      (select released_count + count(id) from ticket where event_id = event.id and status = 'RELEASED' and category_id is null) as released_tickets,
       stats.checked_in_count as checked_in_tickets,
       case(contains_unbounded_categories) when true then
         (select count(id) from ticket where event_id = event.id and status not in ('INVALIDATED', 'EXPIRED'))
           - allocated_count
+          - released_count
           - sold_tickets_count_unbounded
           - checked_in_count_unbounded
           - pending_count_unbounded
@@ -46,6 +96,7 @@ from
 	sum(sold_tickets_count) as sold_tickets_count,
 	sum(checked_in_count) as checked_in_count,
 	sum(pending_count) as pending_count,
+	sum(released_count) as released_count,
 	sum(case (bounded) when true then checked_in_count else 0 end) as checked_in_count_bounded,
 	sum(case (bounded = false) when true then checked_in_count else 0 end) as checked_in_count_unbounded,
 	sum(case (bounded) when true then max_tickets else 0 end) as allocated_count,
