+++ src/main/resources/alfio/db/MYSQL/R__EVENTS_STATISTICS_VIEW.sql
@@ -18,11 +18,70 @@
 drop view if exists events_statistics;
 drop view if exists events_statistics_aggregation_view;
 
+drop view if exists ticket_categories_statistics_pre;
+drop view if exists ticket_category_statistics;
+
+drop view if exists tickets_stats_view;
+drop view if exists tickets_stuck_view;
+
+drop view if exists ticket_category_metainfo_view;
+
+create view ticket_category_metainfo_view as (select max_tickets, bounded, id, event_id, expiration < now() as is_expired from ticket_category);
+
+create view tickets_stats_view as (
+select
+  sum(case(status = 'PENDING') when true then 1 else 0 end) as pending_count,
+  sum(case(status = 'CHECKED_IN') when true then 1 else 0 end) checked_in_count,
+  sum(case(status in ('TO_BE_PAID', 'ACQUIRED')) when true then 1 else 0 end) as sold_tickets_count,
+  sum(case(status = 'RELEASED') when true then 1 else 0 end) as released_count,
+  category_id
+from ticket
+inner join ticket_category tc
+on category_id = tc.id
+group by category_id
+);
+
+create view tickets_stuck_view as (
+select count(*) stuck_count, category_id
+  from ticket
+  inner join tickets_reservation on tickets_reservation.id = tickets_reservation_id
+  where tickets_reservation.status = 'STUCK'
+  group by category_id
+);
+
+create view ticket_categories_statistics_pre as (
+
+select
+  id as ticket_category_id,
+  max_tickets,
+  bounded,
+  is_expired,
+  event_id,
+  coalesce(pending_count,0) as pending_count,
+  coalesce(checked_in_count,0) as checked_in_count,
+  coalesce(sold_tickets_count,0) as sold_tickets_count,
+  coalesce(released_count, 0) as released_count,
+  case(bounded) when false then 0 else max_tickets - coalesce(sold_tickets_count,0 )  - coalesce(checked_in_count, 0) -  coalesce(pending_count, 0) end as not_sold_tickets,
+  coalesce(stuck_count, 0) as stuck_count
+from
+ ticket_category_metainfo_view ticket_cat
+left join tickets_stats_view on ticket_cat.id = tickets_stats_view.category_id
+left join tickets_stuck_view on ticket_cat.id = tickets_stuck_view.category_id
+
+);
+
+create view ticket_category_statistics as (select
+  res.*,
+  is_expired and not_sold_tickets > 0 as is_containing_orphan_tickets,
+  stuck_count > 0 as is_containing_stuck_tickets
+from ticket_categories_statistics_pre res);
+
 create view events_statistics_aggregation_view as (
 select
 	sum(sold_tickets_count) as sold_tickets_count,
 	sum(checked_in_count) as checked_in_count,
 	sum(pending_count) as pending_count,
+	sum(released_count) as released_count,
 	sum(case (bounded) when true then checked_in_count else 0 end) as checked_in_count_bounded,
 	sum(case (bounded = false) when true then checked_in_count else 0 end) as checked_in_count_unbounded,
 	sum(case (bounded) when true then max_tickets else 0 end) as allocated_count,
